requirements.txt
networkx==3.1
spacy==3.7.2
pytest==7.4.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
transformers==4.35.2
torch==2.3.0
Pillow==10.1.0
